cancel entrants button and sample waitlist button
