+++ Trojan0Project/app/src/main/res/layout/activity_view_final_entrants_event.xml
@@ -90,9 +90,40 @@
         android:textColor="#FFFFFF"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintEnd_toStartOf="@+id/cancelEntrantsButton"
+        app:layout_constraintHorizontal_weight="1.1"
         android:layout_marginTop="16dp"
-        android:layout_marginBottom="30dp"/>
+        android:layout_marginBottom="30dp"
+        android:layout_marginEnd="5dp" />
+
+    <!-- "Cancel Entrants" Button -->
+    <Button
+        android:id="@+id/cancelEntrantsButton"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:text="Cancel Entrants"
+        android:backgroundTint="@color/purple"
+        app:layout_constraintStart_toEndOf="@+id/sendNotificationButton"
+        app:layout_constraintEnd_toStartOf="@+id/sampleWaitlistButton"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintHorizontal_weight="1"
+        android:layout_marginTop="16dp"
+        android:layout_marginBottom="30dp"
+        android:layout_marginStart="5dp" />
+
+    <!-- "Sample Waitlist" Button -->
+    <Button
+        android:id="@+id/sampleWaitlistButton"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:text="Sample Waitlist"
+        app:layout_constraintStart_toEndOf="@+id/cancelEntrantsButton"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintHorizontal_weight="1"
+        android:layout_marginTop="16dp"
+        android:layout_marginBottom="30dp"
+        android:layout_marginStart="10dp"/>
 
     <!-- "End of events" TextView placed below the RecyclerView -->
     <TextView
